requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-bootstrap4==1.1.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
djangorestframework==3.11.0
idna==2.9
Pillow==7.1.2
pytz==2019.3
requests==2.23.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
